Implement dtargs.DateType
This is a factory for creating datetime.date instances from strings.
The class can be passed a format string in it's constructor, this can
be anything datime.strptime accepts - if a custom format is not
specified then it will default to '%Y-%m-%d' which is the ISO8061 date
format.