Implement dtargs.DateTimeType
This is a factory for creating datetime.date instances from strings.
The class can be passed a format string in it's constructor, this can
be anything datime.strptime accepts - if a custom format is not
specified then it will default to '%Y-%m-%dT%H:%M:%SZ' which an ISO8061
compliant date/time format (with UTC timezone).